# src/main/resources/static/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: BookShare API
  description: REST API for the BookShare social reading platform
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthorDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
    ReviewDto:
      type: object
      properties:
        reviewedBy:
          type: string
        comment:
          type: string
        reviewDate:
          type: string
          example: "2026-04-21 10:30 AM"
        rating:
          type: integer
          minimum: 0
          maximum: 5
    BookDto:
      type: object
      properties:
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        pages:
          type: integer
          format: int64
        publicationDate:
          type: string
          format: date
          example: "2024-01-15"
        averageRating:
          type: number
          format: double
        authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorDto'
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        reviews:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ReviewDto'
...
    ReadingProgressDto:
      type: object
      properties:
        bookTitle:
          type: string
        isbn:
          type: string
        login:
          type: string
        pagesRead:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        completed:
          type: boolean
...
    ShelfDto:
      type: object
      properties:
        name:
          type: string
        login:
          type: string
        bookCount:
          type: integer
        defaultShelf:
          type: boolean
    ShelfDetailDto:
      type: object
      properties:
        name:
          type: string
        login:
          type: string
        defaultShelf:
          type: boolean
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookDto'
    ShelfCreateDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
    ConnectionApiDto:
      type: object
      required: [handle, action]
      properties:
        handle:
          type: string
        action:
          type: string
          enum:
            - SEND_FRIEND_REQ
            - REVOKE_FRIEND_REQ
            - ACCEPT_FRIEND_REQ
            - DECLINE_FRIEND_REQ
            - REMOVE_FRIEND
    ReadingChallengeDto:
      type: object
      properties:
        email:
          type: string
        year:
          type: integer
        bookCount:
          type: integer
    ReadingChallengeRequest:
      type: object
      required: [bookCount]
      properties:
        year:
          type: integer
        bookCount:
          type: integer
          minimum: 1
          maximum: 1000
    LoginApiDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        handle:
          type: string
        email:
          type: string
        bio:
          type: string
          nullable: true
        profilePictureUrl:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        active:
          type: boolean
        shelfList:
          type: array
          items:
            $ref: '#/components/schemas/ShelfDto'
        readingProgress:
          type: array
          items:
            $ref: '#/components/schemas/ReadingProgressDto'
        connectionList:
          type: array
          items:
            $ref: '#/components/schemas/LoginApiDto'
        reviewList:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
...
    ErrorResponseDto:
      type: object
      properties:
        globalErrors:
          type: array
          items:
            type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: array
              items:
                type: string
...
paths:
...
  /shelf:
    get:
      tags: [Shelf]
      summary: Get all shelves for the authenticated user
      responses:
        '200':
          description: List of shelves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShelfDto'
    post:
      tags: [Shelf]
      summary: Create a new shelf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShelfCreateDto'
      responses:
        '201':
          description: Shelf created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /shelf/{id}:
    get:
      tags: [Shelf]
      summary: Get shelf details including its books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Shelf detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfDetailDto'
        '403':
          description: Shelf does not belong to the authenticated user
  /profile/{handle}:
    get:
      tags: [Profile]
      summary: Get a user profile by handle
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
        - name: detailed
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginApiDto'
...
  /profile/connect:
    post:
      tags: [Profile]
      summary: Perform a connection action (send, accept, revoke, decline, or remove)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionApiDto'
      responses:
        '200':
          description: Action performed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionApiDto'
  /profile/readingChallenge:
    get:
      tags: [Profile]
      summary: Get the reading challenge for the current year for the authenticated user
      responses:
        '200':
          description: Reading challenge found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingChallengeDto'
        '404':
          description: No reading challenge set for the current year
    post:
      tags: [Profile]
      summary: Create or update the reading challenge for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingChallengeRequest'
      responses:
        '200':
          description: Reading challenge saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingChallengeDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
...